完成 删除文章功能， 文章编辑开发 进行中

+++ server/controllers/article.js
@@ -4,7 +4,8 @@ const {
   category: CategoryModel,
   comment: CommentModel,
   reply: ReplyModel,
-  user: UserModel
+  user: UserModel,
+  sequelize
 } = require('../models')
 
 const { checkAuth } = require('../lib/token')
@@ -105,13 +106,27 @@ module.exports = {
 
   // 删除文章
   async delete(ctx) {
-    const { articleId } = ctx.request.body
-    if (articleId) {
-      await TagModel.destroy({ where: { articleId } })
-      await ArticleModel.destroy({ where: { id: articleId } })
-      ctx.body = { code: 200, message: '成功删除文章' }
-    } else {
-      ctx.body = { code: 403, message: '文章 id 不能为空' }
+    const isAuth = checkAuth(ctx)
+    if (isAuth) {
+      const { articleId } = ctx.query
+      if (articleId) {
+        await TagModel.destroy({ where: { articleId } })
+        await ArticleModel.destroy({ where: { id: articleId } })
+        await sequelize.query(
+          // `
+          //   delete article, tag, category, comment, reply from article
+          //   inner join tag on article.id=tag.articleId
+          //   inner join category on article.id=category.articleId
+          //   inner join comment on article.id=comment.articleId
+          //   inner join reply on comment.id=reply.commentId
+          //   where article.id=${articleId}
+          // `
+          `delete comment, reply from comment left join reply on comment.id=reply.commentId where comment.articleId=${articleId}`
+        )
+        ctx.body = { code: 200, message: '成功删除文章' }
+      } else {
+        ctx.body = { code: 403, message: '文章 id 不能为空' }
+      }
     }
   }
 }
